feat: add native tensor foundation and fused execution optimizations
Extend native AST/DAG, Typed IR, planning and worker transport with rank-3
results, named outputs, immutable ModelPayload and bounded iteration.

Fuse pointwise arithmetic into bounded AOT/SIMD regions, reuse lifetime
slots and iteration buffers, write roots into independently owned results,
and precompute immutable short-chain execution metadata. Preserve operator
contracts, input validation, failure isolation and separate multiply/add
rounding. Keep Python as the interface layer.

Add focused regressions, platform computation inventory, mathematical
implementation rules and reproducible before/after/NJIT time and memory
evidence, including unsuccessful candidates and outstanding gates.

Verification of the exact non-Excel candidate on macOS arm64:
- Python: 5106 passed; native: 8/8; ASan/UBSan: 8/8 (LSan not run).
- Native source hashes match the previously benchmarked cf183023 build.
- 118 relative documentation links verified; CSV retains standard CRLF.
- Performance is incomplete: foundation joint gate 1/7, phase-2 gate 2/4.
  Existing thresholds remain unchanged; this is a development checkpoint,
  not migration or release acceptance.

Exclude the parallel Excel implementation, dedicated documents and Excel
sections in shared guidance; leave that work intact in the working tree.

+++ cpp/operators/state_events.cpp
@@ -391,7 +391,7 @@ void prepare_state_events(Prepared &p) {
         if (!p.payload(i)) return false;
         if (valid) return true;
         if (!p.structure_only) throw Error("INVALID_PARAMETER");
-        p.payload_available_mask &= static_cast<std::uint8_t>(~(1u << i));
+        p.payload_available_mask &= static_cast<std::uint32_t>(~(1u << i));
         return false;
     };
     const auto parameter = [&](std::size_t i) {
